:recycle: [FixCircularDependencies] Enhanced command
